feat - like and Member
좋아요 기능과 회원 수정 삭제 설정 추가

+++ src/main/java/com/codestates/server/member/service/MemberService.java
@@ -5,7 +5,12 @@ import com.codestates.server.exception.BusinessLogicException;
 import com.codestates.server.exception.ExceptionCode;
 import com.codestates.server.member.entity.Member;
 import com.codestates.server.member.repository.MemberRepository;
+import com.codestates.server.movie.entity.Movie;
+import com.codestates.server.movie.service.MovieService;
+import com.codestates.server.review.entity.Review;
+import com.codestates.server.review.service.ReviewService;
 import lombok.RequiredArgsConstructor;
+import org.json.simple.parser.ParseException;
 import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -17,6 +22,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
@@ -37,6 +43,8 @@ public class MemberService {
 
     public Member createMember(Member member) {
         // 이미 등록된 이메일인지 확인
+        verifyExistsEmail(member.getEmail());
+
         String encryptedPassword = passwordEncoder.encode(member.getPassword());
         member.setPassword(encryptedPassword);
 
@@ -49,9 +57,13 @@ public class MemberService {
     @Transactional(propagation = Propagation.REQUIRED)
     public Member updateMember(Member member) {
         Member findMember = findVerifiedMember(member.getMemberId());
+        if (findMember.getMemberId() == authenticationMember().getMemberId()) {
+            Optional.ofNullable(member.getUsername())
+                    .ifPresent(name -> findMember.setUsername(name));
+            Optional.ofNullable(member.getPassword())
+                    .ifPresent(password -> findMember.setPassword(passwordEncoder.encode(password)));
+        } else throw new BusinessLogicException(ExceptionCode.MEMBER_FORBIDDEN);
 
-        Optional.ofNullable(member.getUsername())
-                .ifPresent(name -> findMember.setUsername(name));
 
         return memberRepository.save(findMember);
     }
@@ -68,8 +80,9 @@ public class MemberService {
 
     public void deleteMember(long memberId) {
         Member findMember = findVerifiedMember(memberId);
-
-        memberRepository.delete(findMember);
+        if (findMember.getMemberId() == authenticationMember().getMemberId()) {
+            memberRepository.delete(findMember);
+        } else throw new BusinessLogicException(ExceptionCode.MEMBER_FORBIDDEN);
     }
 
     @Transactional(readOnly = true)
